Fold DNSTaskStarter into JmDNSImpl
